and now pack the specialized PrimitiveArrays into the same file for search-and-replace-ability
