+++ src/main/java/edu/vanderbilt/accre/array/JaggedArray.java
@@ -4,10 +4,10 @@ import java.lang.UnsupportedOperationException;
 
 import edu.vanderbilt.accre.interpretation.Interpretation;
 import edu.vanderbilt.accre.array.Array;
-import edu.vanderbilt.accre.array.PrimitiveArrayInt4;
+import edu.vanderbilt.accre.array.PrimitiveArray;
 
 public class JaggedArray extends Array {
-    public JaggedArray(Interpretation interpretation, int length, PrimitiveArrayInt4 offsets, Array content) {
+    public JaggedArray(Interpretation interpretation, int length, PrimitiveArray.Int4 offsets, Array content) {
         super(interpretation, length);
     }
