+++ src/main/java/edu/vanderbilt/accre/array/PrimitiveArray.java
@@ -2,11 +2,14 @@ package edu.vanderbilt.accre.array;
 
 import java.lang.Integer;
 import java.nio.ByteBuffer;
+import java.nio.ByteOrder;
+import java.nio.DoubleBuffer;
+import java.nio.IntBuffer;
 
-import edu.vanderbilt.accre.interpretation.Interpretation;
-import edu.vanderbilt.accre.interpretation.AsDtype;
 import edu.vanderbilt.accre.array.Array;
 import edu.vanderbilt.accre.array.RawArray;
+import edu.vanderbilt.accre.interpretation.AsDtype;
+import edu.vanderbilt.accre.interpretation.Interpretation;
 
 public abstract class PrimitiveArray extends Array {
     ByteBuffer buffer;
@@ -66,4 +69,75 @@ public abstract class PrimitiveArray extends Array {
     }
 
     abstract protected Array make(ByteBuffer out);
+
+    /////////////////////////////////////////////////////////////////////////// Int4
+
+    public static class Int4 extends PrimitiveArray {
+        public Int4(Interpretation interpretation, int length) {
+            super(interpretation, length);
+        }
+
+        public Int4(Interpretation interpretation, RawArray rawarray) {
+            super(interpretation, rawarray);
+        }
+
+        protected Int4(Interpretation interpretation, ByteBuffer buffer) {
+            super(interpretation, buffer);
+        }
+
+        public Int4(int[] data, boolean bigEndian) {
+            super(new AsDtype(AsDtype.Dtype.INT4), data.length);
+            this.buffer = ByteBuffer.allocate(data.length * this.itemsize());
+            this.buffer.order(bigEndian ? ByteOrder.BIG_ENDIAN : ByteOrder.LITTLE_ENDIAN);
+            this.buffer.asIntBuffer().put(data, 0, data.length);
+        }
+
+        public Object toArray(boolean bigEndian) {
+            this.buffer.order(bigEndian ? ByteOrder.BIG_ENDIAN : ByteOrder.LITTLE_ENDIAN);
+            IntBuffer buf = this.buffer.asIntBuffer();
+            int[] out = new int[buf.limit() - buf.position()];
+            buf.get(out);
+            return out;
+        }
+    
+        protected Array make(ByteBuffer out) {
+            return new Int4(this.interpretation, out);
+        }
+    }
+
+    /////////////////////////////////////////////////////////////////////////// Float8
+
+    public static class Float8 extends PrimitiveArray {
+        public Float8(Interpretation interpretation, int length) {
+            super(interpretation, length);
+        }
+
+        public Float8(Interpretation interpretation, RawArray rawarray) {
+            super(interpretation, rawarray);
+        }
+
+        protected Float8(Interpretation interpretation, ByteBuffer buffer) {
+            super(interpretation, buffer);
+        }
+
+        public Float8(double[] data, boolean bigEndian) {
+            super(new AsDtype(AsDtype.Dtype.FLOAT8), data.length);
+            this.buffer = ByteBuffer.allocate(data.length * this.itemsize());
+            this.buffer.order(bigEndian ? ByteOrder.BIG_ENDIAN : ByteOrder.LITTLE_ENDIAN);
+            this.buffer.asDoubleBuffer().put(data, 0, data.length);
+        }
+
+        public Object toArray(boolean bigEndian) {
+            this.buffer.order(bigEndian ? ByteOrder.BIG_ENDIAN : ByteOrder.LITTLE_ENDIAN);
+            DoubleBuffer buf = this.buffer.asDoubleBuffer();
+            double[] out = new double[buf.limit() - buf.position()];
+            buf.get(out);
+            return out;
+        }
+
+        protected Array make(ByteBuffer out) {
+            return new Float8(this.interpretation, out);
+        }
+    }
+    
 }
